Rebased TableEditor on zustand
