+++ src/components/Controllers/Table/Editor.tsx
@@ -7,7 +7,6 @@ export default function Editor() {
   const report = useStore((state) => state.file?.record?.report)
   const loadTable = useStore((state) => state.loadTable)
   const updatePatch = useStore((state) => state.updatePatch)
-  const selectedColumn = useStore((state) => state.selectedColumn)
   if (!table) return null
   if (!report) return null
   return (
@@ -16,7 +15,6 @@ export default function Editor() {
       schema={table.tableSchema}
       report={report}
       onChange={updatePatch}
-      selectedColumn={selectedColumn}
     />
   )
 }
