+++ src/components/Editors/Table/Layout.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react'
+import Box from '@mui/material/Box'
+import Datagrid from './Datagrid'
+
+export default function Layout() {
+  return (
+    <Box height="100%">
+      <Datagrid />
+    </Box>
+  )
+}
