+++ src/components/Editors/Table/interfaces.ts
@@ -1,0 +1,8 @@
+import { IError, IDict } from '../../../interfaces'
+
+export interface IErrorIndex {
+  header: IDict<IError[]>
+  label: IDict<IError[]>
+  row: IDict<IError[]>
+  cell: IDict<IError[]>
+}
