+++ src/components/Editors/Table/settings.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_ACTIVE_CELL: [number, number] = [0, 1]
